+++ src/model_single_phase.jl
@@ -217,6 +217,8 @@ end
 - set net injections Pj/Qj to negative of Inputs.Pload/Qload, which are normalized by Sbase when creating Inputs
 - keys of P/Qload must match Inputs.busses. Any missing keys have load set to zero.
 - Inputs.substation_bus is unconstrained, slack bus
+
+TODO this method is same as LinDistFlow single phase: should it be moved to CommonOPF?
 """
 function constrain_loads(m, p::Inputs)
     Pj = m[:Pj]